Move CanonicalIngredientsAlchemicalProperty model into the Canonical namespace
